Move from fixed configuration to AP config
As configuring the sensor by changing the firmware is not convenient a
wifi access point is available to set the configuration. It is saved to
flash when setting it, but also saved to RTC memory when going to deep
sleep. This allows for a fast reloading on boot and prevents writing the
flash to often.